requirements.txt
celery
attrs
cattrs
